Rename all occurences of "alpha" to "string" in the context of indexes.
Originally, I had the idea that alpha would be short for alphanumeric.
However, over time I've found this name to be confusing. Since it is really
referring to strings (which implicitly contain alphanumeric symbols and other
symbols), that's what I named it now. This is all internal and has no effect on
the public API.