Change Model interface method from GetId to Id.
The new version follows go convention for getter methods.
I also had to change the Id field in DefaultData to be
unexported (now called "id"). This means users who were
previously accessing the Id field directly will need to
use the method. The new version is a little more clear since
it's obvious that a Model needs to have an Id method and not
quite as obvious where the mysterious Id field was coming from.