Use the string "-" instead of "Id" to identify model ids in lua script responses.
This avoids possible collisions. Previously, if there was a model with a field named
"Id", it would not be saved/retrieved correctly because zoom would overwrite that field
with the model id. We could use something like "$id" but that runs the risk of collision
if a custom name for the field was specified via the struct tag `redis:"$id"`. The string
"-" will never cause a collision because the struct tag `redis:"-"` means to ignore the
field, i.e. not store it in the database.