Add config.Password. If provided, connections will authenticate with the AUTH command.
The old way to do this was to provide an address of the form redis://user:pass@localhost:123.
After giving it some thought, I think the new version is better for two reasons: 1) Redis
does not use usernames for authentication so that portion was simply ignored. 2) Password
should be considered a configuration option, so putting it with all the other configuration
options just makes more sense. Documenting the previous address format was also a little more
complicated.