Use a SORT command to get fields for models out of a set of ids.
The old implementation relied on lua scripts, but I've realized that
the same functionality can be acheived with a SORT command that uses
multiple GET options with external hash keys. The new implementation
is more effecient and easier to write code around. Lua scripts will
still be needed, but only for dealing with string indexes.