Implement and test queries with string filters, including combining them with other modifiers.
In order to get this to work, I had to change the format of the string indexes to
value + NULL + id. This now allows string filters to work with one caveat: you cannot
have the NULL character or the DEL character in your strings. I think this is pretty
reasonable, as the most common use case will be filtering ASCII characters, and these
special characters can't be easily typed on a keyboard, inputted into a form, or rendered
on a screen. I'll remember to note this limitation in the README and documentation and
think of ways to possibly improve the implementation in the future.

With this commit, I believe I have reached feature parity with the develop branch. Which means
everything has been rewritten and passes the same tests. I will be merging back into develop
soon, and into master for a new minor point release shortly after that!