Change init function in scripts.go to initScripts. (meaning it will not be run automatically).
initScripts is now called in Init, and since you must call Init in order to start using zoom,
there are no negative effects. The reason I wanted to do this is that with
github.com/soroushjp/humble, we want to be able to share models accross the server and client.
On the client, we don't want to initialize scripts (or use zoom at all). So this change makes it
so that you can use zoom-compatible models with gopherjs without any problems. Zoom will only
attempt to load scripts if you call zoom.Init.

I also reorganized Init and config related code.