Rename Model methods to ModelId and SetModelId. Raname DefaultData to RandomId and export its Id field.
There are two motivations for this change:
1. It is clear at this point that an id is the only thing required of models. Therefore the name DefaultData
does not make any more sense. The name RandomId is more inline with what the embeddable struct actually is
and how it behaves. It also opens the door for other id implementations which may be included out of the box
in the future. e.g., AutoIncrementedId.
2. With an unexported id field, the id was not being included by default when you serialize a model using the
json package (or xml, gob, etc.). This is undesireable since in most cases when you are serializing a model you
want the id to be included. Making it exported fixes the problem, but meant I had to rename the Model interface
methods to avoid conflict. I think ModelId and SetModelId make the most sense here. They are idiomatic (getters
in go typically do not have the "Get" prefix), descriptive, and unlikely to collide with other methods and cause
unexpected behavior.