Remove FindModelsByIdsKey.
This method was too complicated and hard to document. I added it
to Zoom because I thought I needed it for a different project.
However, I ended up using a different approach and this method is
no longer needed. It's extremely unlikely that someone else would
need it, and we can always add it back in later.